+++ src/ext/BalExtension/wixstdba/WixStandardBootstrapperApplication.cpp
@@ -1179,6 +1179,7 @@ private: // privates
     {
         HRESULT hr = S_OK;
         LPWSTR sczLocPath = NULL;
+        LPWSTR sczFormatted = NULL;
         LPCWSTR wzLocFileName = m_fPrereq ? L"mbapreq.wxl" : L"thm.wxl";
 
         hr = LocProbeForFile(wzModulePath, wzLocFileName, wzLanguage, &sczLocPath);
@@ -1198,7 +1199,16 @@ private: // privates
         hr = LocLocalizeString(m_pWixLoc, &m_sczConfirmCloseMessage);
         BalExitOnFailure(hr, "Failed to localize confirm close message: %ls", m_sczConfirmCloseMessage);
 
+        hr = BalFormatString(m_sczConfirmCloseMessage, &sczFormatted);
+        if (SUCCEEDED(hr))
+        {
+            ReleaseStr(m_sczConfirmCloseMessage);
+            m_sczConfirmCloseMessage = sczFormatted;
+            sczFormatted = NULL;
+        }
+
     LExit:
+        ReleaseStr(sczFormatted);
         ReleaseStr(sczLocPath);
 
         return hr;